Use closeTo functions from library.
