+++ package.json
@@ -42,7 +42,7 @@
     "@next/bundle-analyzer": "^10.0.7",
     "@next/mdx": "^9.4.0",
     "@oasisdex/connectors": "^0.0.24",
-    "@oasisdex/multiply": "0.2.5-alpha.1",
+    "@oasisdex/multiply": "0.2.5-alpha.6",
     "@oasisdex/transactions": "^0.0.23",
     "@oasisdex/utils": "^0.0.8",
     "@oasisdex/web3-context": "^0.0.22",
